refactor: remove unused files and models from the project
