--- config/admin.py
@@ -1,8 +1,0 @@
-from django.contrib import admin
-
-from config.models import DefaultImage, Settings, Config
-
-# Register your models here.
-admin.site.register(DefaultImage)
-admin.site.register(Settings)
-admin.site.register(Config)
