--- preview/apps.py
@@ -1,7 +1,0 @@
-from django.apps import AppConfig
-
-
-class PreviewConfig(AppConfig):
-    default_auto_field = 'django.db.models.BigAutoField'
-    name = 'preview'
-    verbose_name = 'preview'
